+++ spec/unit/external_sources_spec.rb
@@ -1,20 +1,37 @@
 require File.expand_path('../../spec_helper', __FILE__)
 
 module Pod
-
   describe ExternalSources do
-    it "returns the instance of appropriate concrete class according to the parameters" do
-      git     = Dependency.new("Reachability", :git     => nil)
-      svn     = Dependency.new("Reachability", :svn     => nil)
-      podspec = Dependency.new("Reachability", :podspec => nil)
-      local   = Dependency.new("Reachability", :local   => nil)
-      path    = Dependency.new("Reachability", :path   => nil)
+    before do
+      @subject = ExternalSources
+    end
 
-      ExternalSources.from_dependency(git, nil).class.should     == ExternalSources::GitSource
-      ExternalSources.from_dependency(svn, nil).class.should     == ExternalSources::SvnSource
-      ExternalSources.from_dependency(podspec, nil).class.should == ExternalSources::PodspecSource
-      ExternalSources.from_dependency(local, nil).class.should   == ExternalSources::PathSource
-      ExternalSources.from_dependency(path, nil).class.should    == ExternalSources::PathSource
+    describe "from_dependency" do
+      it "supports a podspec source" do
+        dep = Dependency.new("Reachability", :podspec => nil)
+        klass = @subject.from_dependency(dep, nil).class
+        klass.should == @subject::PodspecSource
+      end
+
+      it "supports a path source" do
+        dep = Dependency.new("Reachability", :path => nil)
+        klass = @subject.from_dependency(dep, nil).class
+        klass.should == @subject::PathSource
+      end
+
+      it "supports a path source specified with the legacy :local key" do
+        dep = Dependency.new("Reachability", :local => nil)
+        klass = @subject.from_dependency(dep, nil).class
+        klass.should == @subject::PathSource
+      end
+
+      it "supports all the strategies implemented by the downloader" do
+        [:git, :svn, :hg, :bzr, :http].each do |strategy|
+          dep     = Dependency.new("Reachability", strategy => nil)
+          klass = @subject.from_dependency(dep, nil).class
+          klass.should == @subject::DownloaderSource
+        end
+      end
     end
   end
 
@@ -24,7 +41,7 @@ module Pod
 
     before do
       dependency = Dependency.new("Reachability", :git => fixture('integration/Reachability'))
-      @external_source = ExternalSources.from_dependency(dependency, nil)
+      @subject = ExternalSources.from_dependency(dependency, nil)
     end
 
     #--------------------------------------#
@@ -43,7 +60,7 @@ module Pod
 
       it "fetches the specification from the remote stores it in the sandbox" do
         config.sandbox.specification('Reachability').should == nil
-        @external_source.fetch(config.sandbox)
+        @subject.fetch(config.sandbox)
         config.sandbox.specification('Reachability').name.should == 'Reachability'
       end
 
@@ -55,7 +72,7 @@ module Pod
 
       it "pre-downloads the Pod and stores the relevant information in the sandbox" do
         sandbox = config.sandbox
-        @external_source.send(:pre_download, sandbox)
+        @subject.send(:pre_download, sandbox)
         path = config.sandbox.root + 'Local Podspecs/Reachability.podspec'
         path.should.exist?
         sandbox.predownloaded_pods.should == ["Reachability"]
@@ -73,101 +90,31 @@ module Pod
 
   #---------------------------------------------------------------------------#
 
-  describe ExternalSources::GitSource do
+  describe ExternalSources::DownloaderSource do
 
     before do
-      dependency = Dependency.new("Reachability", :git => fixture('integration/Reachability'))
-      @external_source = ExternalSources.from_dependency(dependency, nil)
+      params = {
+        :git => fixture('integration/Reachability'),
+        :branch => 'master'
+      }
+      dep = Dependency.new("Reachability", params)
+      @subject = ExternalSources.from_dependency(dep, nil)
     end
 
     it "creates a copy of the podspec" do
-      @external_source.fetch(config.sandbox)
+      @subject.fetch(config.sandbox)
       path = config.sandbox.root + 'Local Podspecs/Reachability.podspec'
       path.should.exist?
     end
 
-    it "marks a LocalPod as downloaded" do
-      @external_source.fetch(config.sandbox)
+    it "marks the Pod as pre-downloaded" do
+      @subject.fetch(config.sandbox)
       config.sandbox.predownloaded_pods.should == ["Reachability"]
     end
 
     it "returns the description" do
-      @external_source.description.should.match %r|from `.*Reachability`|
-    end
-  end
-
-  #---------------------------------------------------------------------------#
-
-  describe ExternalSources::SvnSource do
-
-    before do
-      dependency = Dependency.new("SvnSource", :svn => "file://#{fixture('subversion-repo/trunk')}")
-      @external_source = ExternalSources.from_dependency(dependency, nil)
-    end
-
-    it "creates a copy of the podspec" do
-      @external_source.fetch(config.sandbox)
-      path = config.sandbox.root + 'Local Podspecs/SvnSource.podspec'
-      path.should.exist?
-    end
-
-    it "marks a LocalPod as downloaded" do
-      @external_source.fetch(config.sandbox)
-      config.sandbox.predownloaded_pods.should == ["SvnSource"]
-    end
-
-    it "returns the description" do
-      @external_source.description.should.match %r|from `.*subversion-repo/trunk`|
-    end
-  end
-
-  #---------------------------------------------------------------------------#
-
-  describe ExternalSources::MercurialSource do
-
-    before do
-      dependency = Dependency.new("MercurialSource", :hg => fixture('mercurial-repo'))
-      @external_source = ExternalSources.from_dependency(dependency, nil)
-    end
-
-    it "creates a copy of the podspec" do
-      @external_source.fetch(config.sandbox)
-      path = config.sandbox.root + 'Local Podspecs/MercurialSource.podspec'
-      path.should.exist?
-    end
-
-    it "marks a LocalPod as downloaded" do
-      @external_source.fetch(config.sandbox)
-      config.sandbox.predownloaded_pods.should == ["MercurialSource"]
-    end
-
-    it "returns the description" do
-      @external_source.description.should.match %r|from `.*/mercurial-repo`|
-    end
-  end
-
-  #---------------------------------------------------------------------------#
-
-  describe ExternalSources::BazaarSource do
-
-    before do
-      dependency = Dependency.new("BazaarSource", :bzr => fixture('bzr-repo'))
-      @external_source = ExternalSources.from_dependency(dependency, nil)
-    end
-
-    it "creates a copy of the podspec" do
-      @external_source.fetch(config.sandbox)
-      path = config.sandbox.root + 'Local Podspecs/BazaarSource.podspec'
-      path.should.exist?
-    end
-
-    it "marks a LocalPod as downloaded" do
-      @external_source.fetch(config.sandbox)
-      config.sandbox.predownloaded_pods.should == ["BazaarSource"]
-    end
-
-    it "returns the description" do
-      @external_source.description.should.match %r|from `.*/bzr-repo`|
+      expected = /from `.*Reachability`, branch `master`/
+      @subject.description.should.match(expected)
     end
   end
 
@@ -179,49 +126,49 @@ module Pod
       podspec_path = fixture('integration/Reachability/Reachability.podspec')
       dependency = Dependency.new("Reachability", :podspec => podspec_path.to_s)
       podfile_path = fixture('integration/Podfile')
-      @external_source = ExternalSources.from_dependency(dependency, podfile_path)
+      @subject = ExternalSources.from_dependency(dependency, podfile_path)
     end
 
     it "creates a copy of the podspec" do
-      @external_source.fetch(config.sandbox)
+      @subject.fetch(config.sandbox)
       path = config.sandbox.root + 'Local Podspecs/Reachability.podspec'
       path.should.exist?
     end
 
     it "returns the description" do
-      @external_source.description.should.match %r|from `.*Reachability/Reachability.podspec`|
+      @subject.description.should.match %r|from `.*Reachability/Reachability.podspec`|
     end
 
     describe "Helpers" do
 
       it "handles absolute paths" do
-        @external_source.stubs(:params).returns(:podspec => fixture('integration/Reachability'))
-        path = @external_source.send(:podspec_uri)
+        @subject.stubs(:params).returns(:podspec => fixture('integration/Reachability'))
+        path = @subject.send(:podspec_uri)
         path.should == fixture('integration/Reachability/Reachability.podspec').to_s
       end
 
       it "handles paths when there is no podfile path" do
-        @external_source.stubs(:podfile_path).returns(nil)
-        @external_source.stubs(:params).returns(:podspec => fixture('integration/Reachability'))
-        path = @external_source.send(:podspec_uri)
+        @subject.stubs(:podfile_path).returns(nil)
+        @subject.stubs(:params).returns(:podspec => fixture('integration/Reachability'))
+        path = @subject.send(:podspec_uri)
         path.should == fixture('integration/Reachability/Reachability.podspec').to_s
       end
 
       it "handles relative paths" do
-        @external_source.stubs(:params).returns(:podspec => 'Reachability')
-        path = @external_source.send(:podspec_uri)
+        @subject.stubs(:params).returns(:podspec => 'Reachability')
+        path = @subject.send(:podspec_uri)
         path.should == fixture('integration/Reachability/Reachability.podspec').to_s
       end
 
       it "expands the tilde" do
-        @external_source.stubs(:params).returns(:podspec => '~/Reachability')
-        path = @external_source.send(:podspec_uri)
+        @subject.stubs(:params).returns(:podspec => '~/Reachability')
+        path = @subject.send(:podspec_uri)
         path.should == ENV['HOME'] + '/Reachability/Reachability.podspec'
       end
 
-      it "handles urls" do
-        @external_source.stubs(:params).returns(:podspec => "http://www.example.com/Reachability.podspec")
-        path = @external_source.send(:podspec_uri)
+      it "handles URLs" do
+        @subject.stubs(:params).returns(:podspec => "http://www.example.com/Reachability.podspec")
+        path = @subject.send(:podspec_uri)
         path.should == "http://www.example.com/Reachability.podspec"
       end
     end
@@ -235,11 +182,11 @@ module Pod
       podspec_path = fixture('integration/Reachability/Reachability.podspec')
       dependency = Dependency.new("Reachability", :path => fixture('integration/Reachability'))
       podfile_path = fixture('integration/Podfile')
-      @external_source = ExternalSources.from_dependency(dependency, podfile_path)
+      @subject = ExternalSources.from_dependency(dependency, podfile_path)
     end
 
     it "creates a copy of the podspec" do
-      @external_source.fetch(config.sandbox)
+      @subject.fetch(config.sandbox)
       path = config.sandbox.root + 'Local Podspecs/Reachability.podspec'
       path.should.exist?
     end
@@ -254,11 +201,11 @@ module Pod
     end
 
     it "returns the description" do
-      @external_source.description.should.match %r|from `.*integration/Reachability`|
+      @subject.description.should.match %r|from `.*integration/Reachability`|
     end
 
     it "marks the Pod as local in the sandbox" do
-      @external_source.fetch(config.sandbox)
+      @subject.fetch(config.sandbox)
       config.sandbox.development_pods.should == {
         "Reachability" => fixture('integration/Reachability').to_s
       }
@@ -267,34 +214,34 @@ module Pod
     describe "Helpers" do
 
       it "handles absolute paths" do
-        @external_source.stubs(:params).returns(:path => fixture('integration/Reachability'))
-        path = @external_source.send(:podspec_path)
+        @subject.stubs(:params).returns(:path => fixture('integration/Reachability'))
+        path = @subject.send(:podspec_path)
         path.should == fixture('integration/Reachability/Reachability.podspec')
       end
 
       it "handles paths when there is no podfile path" do
-        @external_source.stubs(:podfile_path).returns(nil)
-        @external_source.stubs(:params).returns(:path => fixture('integration/Reachability'))
-        path = @external_source.send(:podspec_path)
+        @subject.stubs(:podfile_path).returns(nil)
+        @subject.stubs(:params).returns(:path => fixture('integration/Reachability'))
+        path = @subject.send(:podspec_path)
         path.should == fixture('integration/Reachability/Reachability.podspec')
       end
 
       it "handles relative paths" do
-        @external_source.stubs(:params).returns(:path => 'Reachability')
-        path = @external_source.send(:podspec_path)
+        @subject.stubs(:params).returns(:path => 'Reachability')
+        path = @subject.send(:podspec_path)
         path.should == fixture('integration/Reachability/Reachability.podspec')
       end
 
       it "expands the tilde" do
-        @external_source.stubs(:params).returns(:path => '~/Reachability')
+        @subject.stubs(:params).returns(:path => '~/Reachability')
         Pathname.any_instance.stubs(:exist?).returns(true)
-        path = @external_source.send(:podspec_path)
+        path = @subject.send(:podspec_path)
         path.should == Pathname(ENV['HOME']) + 'Reachability/Reachability.podspec'
       end
 
       it "raises if the podspec cannot be found" do
-        @external_source.stubs(:params).returns(:path => temporary_directory)
-        e = lambda { @external_source.send(:podspec_path) }.should.raise Informative
+        @subject.stubs(:params).returns(:path => temporary_directory)
+        e = lambda { @subject.send(:podspec_path) }.should.raise Informative
         e.message.should.match /No podspec found for `Reachability` in `#{temporary_directory}`/
       end
     end